serviceCataloguePrestations/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
py-eureka-client==0.11.2
django-cors-headers==4.3.1
requests==2.31.0
python-decouple==3.8